+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/ChangeFeedQueryImpl.java
@@ -6,6 +6,7 @@ import com.azure.cosmos.BridgeInternal;
 import com.azure.cosmos.implementation.changefeed.implementation.ChangeFeedState;
 import com.azure.cosmos.implementation.changefeed.implementation.ChangeFeedStateV1;
 import com.azure.cosmos.implementation.feedranges.FeedRangeInternal;
+import com.azure.cosmos.implementation.query.DocumentQueryExecutionContextBase;
 import com.azure.cosmos.implementation.query.Paginator;
 import com.azure.cosmos.implementation.spark.OperationContext;
 import com.azure.cosmos.implementation.spark.OperationContextAndListenerTuple;
@@ -13,6 +14,7 @@ import com.azure.cosmos.implementation.spark.OperationListener;
 import com.azure.cosmos.models.CosmosChangeFeedRequestOptions;
 import com.azure.cosmos.models.FeedResponse;
 import com.azure.cosmos.models.ModelBridgeInternal;
+import com.fasterxml.jackson.databind.JsonNode;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -23,7 +25,7 @@ import java.util.function.Supplier;
 
 import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkNotNull;
 
-class ChangeFeedQueryImpl<T extends Resource> {
+class ChangeFeedQueryImpl<T> {
 
     private static final int INITIAL_TOP_VALUE = -1;
 
@@ -37,6 +39,7 @@ class ChangeFeedQueryImpl<T extends Resource> {
     private final ResourceType resourceType;
     private final ChangeFeedState changeFeedState;
     private final OperationContextAndListenerTuple operationContextAndListener;
+    private final Function<JsonNode, T> factoryMethod;
 
     public ChangeFeedQueryImpl(
         RxDocumentClientImpl client,
@@ -69,6 +72,8 @@ class ChangeFeedQueryImpl<T extends Resource> {
         this.klass = klass;
         this.documentsLink = Utils.joinPath(collectionLink, Paths.DOCUMENTS_PATH_SEGMENT);
         this.options = requestOptions;
+        this.factoryMethod = DocumentQueryExecutionContextBase
+            .getEffectiveFactoryMethod(options, klass);
         this.operationContextAndListener = ImplementationBridgeHelpers
                 .CosmosChangeFeedRequestOptionsHelper
                 .getCosmosChangeFeedRequestOptionsAccessor()
@@ -97,7 +102,6 @@ class ChangeFeedQueryImpl<T extends Resource> {
             ModelBridgeInternal.getPropertiesFromChangeFeedRequestOptions(this.options),
             this.createRequestFunc,
             this.executeFunc,
-            this.klass,
             INITIAL_TOP_VALUE,
             this.options.getMaxItemCount(),
             this.options.getMaxPrefetchPageCount(),
@@ -128,7 +132,7 @@ class ChangeFeedQueryImpl<T extends Resource> {
     private Mono<FeedResponse<T>> executeRequestAsync(RxDocumentServiceRequest request) {
         if (this.operationContextAndListener == null) {
             return client.readFeed(request)
-                         .map(rsp -> BridgeInternal.toChangeFeedResponsePage(rsp, klass));
+                         .map(rsp -> BridgeInternal.toChangeFeedResponsePage(rsp, this.factoryMethod, klass));
         } else {
             final OperationListener listener = operationContextAndListener.getOperationListener();
             final OperationContext operationContext = operationContextAndListener.getOperationContext();
@@ -141,7 +145,8 @@ class ChangeFeedQueryImpl<T extends Resource> {
                          .map(rsp -> {
                              listener.responseListener(operationContext, rsp);
 
-                             final FeedResponse<T> feedResponse = BridgeInternal.toChangeFeedResponsePage(rsp, klass);
+                             final FeedResponse<T> feedResponse = BridgeInternal.toChangeFeedResponsePage(
+                                 rsp, this.factoryMethod, klass);
                              listener.feedResponseReceivedListener(operationContext, feedResponse);
 
                              return feedResponse;