+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosChangeFeedRequestOptions.java
@@ -13,10 +13,12 @@ import com.azure.cosmos.implementation.feedranges.FeedRangeInternal;
 import com.azure.cosmos.implementation.query.CompositeContinuationToken;
 import com.azure.cosmos.implementation.spark.OperationContextAndListenerTuple;
 import com.azure.cosmos.util.Beta;
+import com.fasterxml.jackson.databind.JsonNode;
 
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.function.Function;
 
 import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkArgument;
 import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkNotNull;
@@ -41,6 +43,7 @@ public final class CosmosChangeFeedRequestOptions {
     private String throughputControlGroupName;
     private Map<String, String> customOptions;
     private OperationContextAndListenerTuple operationContextAndListenerTuple;
+    private Function<JsonNode, ?> itemFactoryMethod;
 
     private CosmosChangeFeedRequestOptions(
         FeedRangeInternal feedRange,
@@ -478,6 +481,13 @@ public final class CosmosChangeFeedRequestOptions {
         return this.operationContextAndListenerTuple;
     }
 
+    Function<JsonNode, ?> getItemFactoryMethod() { return this.itemFactoryMethod; }
+
+    CosmosChangeFeedRequestOptions setItemFactoryMethod(Function<JsonNode, ?> factoryMethod) {
+        this.itemFactoryMethod = factoryMethod;
+        return this;
+    }
+
     ///////////////////////////////////////////////////////////////////////////////////////////
     // the following helper/accessor only helps to access this class outside of this package.//
     ///////////////////////////////////////////////////////////////////////////////////////////
@@ -513,6 +523,22 @@ public final class CosmosChangeFeedRequestOptions {
                 ) {
 
                     return changeFeedRequestOptions.getOperationContextAndListenerTuple();
+                }
+
+                @Override
+                @SuppressWarnings("unchecked")
+                public <T> Function<JsonNode, T> getItemFactoryMethod(
+                    CosmosChangeFeedRequestOptions options, Class<T> classOfT) {
+
+                    return (Function<JsonNode, T>)options.getItemFactoryMethod();
+                }
+
+                @Override
+                public CosmosChangeFeedRequestOptions setItemFactoryMethod(
+                    CosmosChangeFeedRequestOptions options,
+                    Function<JsonNode, ?> factoryMethod) {
+
+                    return options.setItemFactoryMethod(factoryMethod);
                 }
             });
     }