+++ sdk/schemaregistry/azure-data-schemaregistry-apacheavro/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Release History
 
-## 1.0.0-beta.10 (Unreleased)
+## 1.0.0-beta.11 (Unreleased)
 
 ### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+- Fixed a bug that caused deserialize operation to throw `SchemaParseException` when multiple messages with same schema 
+  were deserialized (https://github.com/Azure/azure-sdk-for-java/issues/27602).
+
+### Other Changes
+
+## 1.0.0-beta.10 (2022-03-15)
 
 ### Breaking Changes
 
@@ -10,9 +20,11 @@
 - Changed `decodeMessageData` and `decodeMessageDataAsync` to `deserializeMessageData` and `deserializeMessageDataAsync`.
 - Changed `encodeMessageData` and `encodeMessageDataAsync` to `serializeMessageData` and `serializeMessageDataAsync`.
 
-### Bugs Fixed
-
 ### Other Changes
+
+#### Dependency Updates
+
+- Update `azure-data-schemaregistry` dependency to `1.0.3`.
 
 ## 1.0.0-beta.9 (2022-02-12)